+++ tests/testthat/test_exact_extract.R
@@ -854,3 +854,32 @@ test_that('Error is thrown if quantiles not specified or not valid', {
  expect_error(exact_extract(rast, square, 'quantile', quantiles=numeric()),
               'Quantiles not specified')
 })
+
+test_that('Progress bar updates incrementally', {
+  rast <- make_square_raster(1:100)
+
+  npolys <- 13
+
+  polys <- st_sf(fid = seq_len(npolys),
+                 geometry = st_sfc(replicate(npolys, {
+                   x <- runif(1, min=0, max=10)
+                   y <- runif(1, min=0, max=10)
+                   r <- runif(1, min=0, max=2)
+                   make_circle(x, y, r, crs=sf::st_crs(rast))
+                 }), crs=sf::st_crs(rast)))
+
+  for (fun in list('sum', weighted.mean)) {
+    for (input in list(polys, sf::st_geometry(polys))) {
+      output <- capture.output(q <- exact_extract(rast, input, fun))
+      lines <- strsplit(output, '\r', fixed=TRUE)[[1]]
+      numlines <- lines[endsWith(lines, '%')]
+      len <- nchar(numlines[1])
+      pcts <- as.integer(substr(numlines, len - 3, len - 1))
+
+      expect_length(pcts, 1 + npolys)
+      expect_equal(pcts[1], 0)
+      expect_equal(pcts[length(pcts)], 100)
+      expect_false(is.unsorted(pcts))
+    }
+  }
+})