+++ NEWS.md
@@ -2,6 +2,8 @@
 
 - Add exact_resample
 - Add force_df and full_colnames arguments to exact_extract, to obtain output in a consistent format
+- Add include_cols and append_cols arguments to exact_extract for linking pixels values or  summarized 
+  results to input features
 - Add stack_apply argument to exact_extract to apply an R function individually to each layer in
   a RasterStack / RasterBrick.
 - Support returning a data frame from an R function applied to each polygon, then combining those